# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Message Manager Server.",
        "title": "Message Manager",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/message_center/config/push": {
            "get": {
                "description": "get push config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_push"
                ],
                "summary": "GetPushConfig",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/app.MessagePushDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "update a push_config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_push"
                ],
                "summary": "UpdatePushConfig",
                "parameters": [
                    {
                        "description": "updatePushConfigDTO",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.updatePushConfigDTO"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "add a new push_config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_push"
                ],
                "summary": "AddPushConfig",
                "parameters": [
                    {
                        "description": "newPushConfigDTO",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.newPushConfigDTO"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a push_config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_push"
                ],
                "summary": "RemovePushConfig",
                "parameters": [
                    {
                        "description": "deletePushConfigDTO",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.deletePushConfigDTO"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/message_center/inner": {
            "get": {
                "description": "get inner message",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_center"
                ],
                "summary": "GetInnerMessage",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/app.MessageListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "set message read",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_center"
                ],
                "summary": "SetMessageIsRead",
                "parameters": [
                    {
                        "description": "messageStatus",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.messageStatus"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "remove message",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_center"
                ],
                "summary": "RemoveMessage",
                "parameters": [
                    {
                        "description": "messageStatus",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.messageStatus"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/message_center/inner_quick": {
            "get": {
                "description": "get inner message by filter",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message_center"
                ],
                "summary": "GetInnerMessageQuick",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/app.MessageListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.MessageListDTO": {
            "type": "object",
            "properties": {
                "data_content_type": {
                    "type": "string"
                },
                "data_schema": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "source": {
                    "type": "string"
                },
                "source_group": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "spec_version": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "app.MessagePushDTO": {
            "type": "object",
            "properties": {
                "mail": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "need_inner_message": {
                    "type": "boolean"
                },
                "need_mail": {
                    "type": "boolean"
                },
                "need_message": {
                    "type": "boolean"
                },
                "need_phone": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "recipient_name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "subscribe_id": {
                    "type": "integer"
                }
            }
        },
...
        "controller.deletePushConfigDTO": {
            "type": "object",
            "properties": {
                "recipient_id": {
                    "type": "integer"
                },
                "subscribe_id": {
                    "type": "integer"
                }
            }
        },
...
        "controller.messageStatus": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "controller.newPushConfigDTO": {
            "type": "object",
            "properties": {
                "need_inner_message": {
                    "type": "boolean"
                },
                "need_mail": {
                    "type": "boolean"
                },
                "need_message": {
                    "type": "boolean"
                },
                "need_phone": {
                    "type": "boolean"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "subscribe_id": {
                    "type": "integer"
                }
            }
        },
...
        "controller.updatePushConfigDTO": {
            "type": "object",
            "properties": {
                "need_inner_message": {
                    "type": "boolean"
                },
                "need_mail": {
                    "type": "boolean"
                },
                "need_message": {
                    "type": "boolean"
                },
                "need_phone": {
                    "type": "boolean"
                },
                "recipient_id": {
                    "type": "string"
                },
                "subscribe_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}